# codegen/backend-services/backend.swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "HotelManagement.WebApi",
    "version": "1.0"
  },
  "paths": {
...
    "/properties": {
      "get": {
        "tags": [
          "Property"
        ],
        "parameters": [
          {
            "name": "from",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "to",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PropertySummaryIPaginatedResult"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "post": {
        "tags": [
          "Property"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePropertyCommand"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePropertyCommand"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePropertyCommand"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          }
        }
      },
      "patch": {
        "tags": [
          "Property"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePropertyCommand"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePropertyCommand"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePropertyCommand"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/properties/filter/{location}/{startDate}/{endDate}/{numberOfAdults}/{numberOfChildren}/{numberOfRooms}": {
      "get": {
        "tags": [
          "Property"
        ],
        "parameters": [
          {
            "name": "from",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "to",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "location",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "startDate",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "endDate",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "numberOfAdults",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "numberOfChildren",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "numberOfRooms",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "LoggedUserId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "HasFreeWiFi",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "HasParking",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "HasRoomService",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "HasRestaurant",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "HasBreakfast",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "HasKitchen",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "HasPool",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "HasFitnessCenter",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "HasPetFriendlyPolicy",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "HasFreeCancellation",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "HasPrivateBathroom",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "HasAirConditioning",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "HasTowels",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "HasHairdryer",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "HasBalcony",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "HasSeaView",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "HasRefrigerator",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "RatingOver9",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "RatingOver8",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "RatingOver7",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "RatingOver6",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PropertySummaryFilteredIPaginatedResult"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/properties/recommendations": {
      "get": {
        "tags": [
          "Property"
        ],
        "parameters": [
          {
            "name": "from",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "to",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "logggedUsedId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "searchHistoryFields",
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/SearchHistoryFields"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PropertySummaryRecommendationIPaginatedResult"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/properties/types": {
      "get": {
        "tags": [
          "Property"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PropertyTypeSummary"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/properties/{id}/reviews": {
      "get": {
        "tags": [
          "Property"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "from",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "to",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PropertyReviewIPaginatedResult"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/properties/{id}/rooms": {
      "get": {
        "tags": [
          "Property"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "from",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "to",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "startDate",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "endDate",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "numberOfAdults",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "numberOfChildren",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "loggedUserId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PropertyRoomIPaginatedResult"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/properties/{id}": {
      "get": {
        "tags": [
          "Property"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "startDate",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "endDate",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "numberOfAdults",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "numberOfChildren",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "loggedUserId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PropertyDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      },
      "delete": {
        "tags": [
          "Property"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/properties/discount": {
      "get": {
        "tags": [
          "Property"
        ],
        "parameters": [
          {
            "name": "propertyId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "loggedUserId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/properties/upload": {
      "patch": {
        "tags": [
          "Property"
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "File": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              },
              "encoding": {
                "File": {
                  "style": "form"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreatePropertyCommand": {
        "required": [
          "description",
          "email",
          "hasBreakfast",
          "hasFitnessCenter",
          "hasFreeCancellation",
          "hasFreeWiFi",
          "hasKitchen",
          "hasParking",
          "hasPetFriendlyPolicy",
          "hasPool",
          "hasRestaurant",
          "hasRoomService",
          "location",
          "name",
          "phoneNumber",
          "pictureUrls",
          "prepaymentNeeded",
          "rating",
          "type"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "type": {
            "$ref": "#/components/schemas/PropertyType"
          },
          "description": {
            "type": "string"
          },
          "location": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "phoneNumber": {
            "type": "string"
          },
          "rating": {
            "type": "integer",
            "format": "int32"
          },
          "prepaymentNeeded": {
            "type": "boolean"
          },
          "hasFreeWiFi": {
            "type": "boolean"
          },
          "hasParking": {
            "type": "boolean"
          },
          "hasKitchen": {
            "type": "boolean"
          },
          "hasPool": {
            "type": "boolean"
          },
          "hasRestaurant": {
            "type": "boolean"
          },
          "hasFitnessCenter": {
            "type": "boolean"
          },
          "hasRoomService": {
            "type": "boolean"
          },
          "hasPetFriendlyPolicy": {
            "type": "boolean"
          },
          "hasBreakfast": {
            "type": "boolean"
          },
          "hasFreeCancellation": {
            "type": "boolean"
          },
          "pictureUrls": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
...
      "PropertyReview": {
        "required": [
          "createdOn",
          "description",
          "id",
          "profilePicture",
          "rating",
          "rowNumber",
          "title",
          "userFirstName",
          "userLastName",
          "userNationality"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "rating": {
            "type": "number",
            "format": "double"
          },
          "createdOn": {
            "type": "string",
            "format": "date-time"
          },
          "userFirstName": {
            "type": "string"
          },
          "userLastName": {
            "type": "string"
          },
          "userNationality": {
            "type": "string"
          },
          "profilePicture": {
            "type": "string"
          },
          "rowNumber": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "PropertyReviewIPaginatedResult": {
        "required": [
          "results",
          "totalCount"
        ],
        "type": "object",
        "properties": {
          "results": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PropertyReview"
            },
            "readOnly": true
          },
          "totalCount": {
            "type": "integer",
            "format": "int32",
            "readOnly": true
          }
        },
        "additionalProperties": false
      },
...
      "PropertyRoomIPaginatedResult": {
        "required": [
          "results",
          "totalCount"
        ],
        "type": "object",
        "properties": {
          "results": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PropertyRoom"
            },
            "readOnly": true
          },
          "totalCount": {
            "type": "integer",
            "format": "int32",
            "readOnly": true
          }
        },
        "additionalProperties": false
      },
      "PropertySummary": {
        "required": [
          "createdOn",
          "id",
          "location",
          "name",
          "numberOfRooms",
          "rowNumber"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "location": {
            "type": "string"
          },
          "createdOn": {
            "type": "string",
            "format": "date-time"
          },
          "numberOfRooms": {
            "type": "integer",
            "format": "int32"
          },
          "rowNumber": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
      "PropertySummaryFilteredIPaginatedResult": {
        "required": [
          "results",
          "totalCount"
        ],
        "type": "object",
        "properties": {
          "results": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PropertySummaryFiltered"
            },
            "readOnly": true
          },
          "totalCount": {
            "type": "integer",
            "format": "int32",
            "readOnly": true
          }
        },
        "additionalProperties": false
      },
      "PropertySummaryIPaginatedResult": {
        "required": [
          "results",
          "totalCount"
        ],
        "type": "object",
        "properties": {
          "results": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PropertySummary"
            },
            "readOnly": true
          },
          "totalCount": {
            "type": "integer",
            "format": "int32",
            "readOnly": true
          }
        },
        "additionalProperties": false
      },
...
      "PropertySummaryRecommendationIPaginatedResult": {
        "required": [
          "results",
          "totalCount"
        ],
        "type": "object",
        "properties": {
          "results": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PropertySummaryRecommendation"
            },
            "readOnly": true
          },
          "totalCount": {
            "type": "integer",
            "format": "int32",
            "readOnly": true
          }
        },
        "additionalProperties": false
      },
      "PropertyType": {
        "enum": [
          0,
          1,
          2,
          3
        ],
        "type": "integer",
        "format": "int32"
      },
      "PropertyTypeSummary": {
        "required": [
          "type",
          "value"
        ],
        "type": "object",
        "properties": {
          "type": {
            "type": "string"
          },
          "value": {
            "$ref": "#/components/schemas/PropertyType"
          }
        },
        "additionalProperties": false
      },
...
      "SearchHistoryFields": {
        "required": [
          "location",
          "numberOfAdults",
          "numberOfChildren",
          "numberOfRooms"
        ],
        "type": "object",
        "properties": {
          "location": {
            "type": "string"
          },
          "numberOfAdults": {
            "type": "integer",
            "format": "int32"
          },
          "numberOfChildren": {
            "type": "integer",
            "format": "int32"
          },
          "numberOfRooms": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
      "UpdatePropertyCommand": {
        "required": [
          "description",
          "email",
          "hasBreakfast",
          "hasFitnessCenter",
          "hasFreeCancellation",
          "hasFreeWiFi",
          "hasKitchen",
          "hasParking",
          "hasPetFriendlyPolicy",
          "hasPool",
          "hasRestaurant",
          "hasRoomService",
          "id",
          "imageUrls",
          "name",
          "phoneNumber",
          "prepaymentNeeded",
          "rating"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "phoneNumber": {
            "type": "string"
          },
          "rating": {
            "type": "integer",
            "format": "int32"
          },
          "prepaymentNeeded": {
            "type": "boolean"
          },
          "hasFreeWiFi": {
            "type": "boolean"
          },
          "hasParking": {
            "type": "boolean"
          },
          "hasKitchen": {
            "type": "boolean"
          },
          "hasPool": {
            "type": "boolean"
          },
          "hasRestaurant": {
            "type": "boolean"
          },
          "hasFitnessCenter": {
            "type": "boolean"
          },
          "hasRoomService": {
            "type": "boolean"
          },
          "hasPetFriendlyPolicy": {
            "type": "boolean"
          },
          "hasBreakfast": {
            "type": "boolean"
          },
          "hasFreeCancellation": {
            "type": "boolean"
          },
          "imageUrls": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}